requirements.txt
Flask==2.2.5
openai==1.8.0
PyGithub==2.1.1
python-dotenv==1.0.1
